refactor: change movie card naming
