+++ src/components/MovieCard/MovieCard.js
@@ -1,4 +1,4 @@
-import "./MoviesCard.css";
+import "./MovieCard.css";
 import { MOVIES_BASE_URL } from '../../utils/Constants';
 import { formatTime } from "../../utils/Utils";
 import { useState } from "react";
@@ -10,7 +10,7 @@ function MoviesCard({ movie }) {
   const { pathname } = useLocation();
 
   const saveButtonClassName = ( 
-    `movies-card__button movies-card__button_type_save ${isSaved && 'movies-card__button_saved'}` 
+    `movie-card__button movie-card__button_type_save ${isSaved && 'movie-card__button_saved'}` 
   );
 
   function handleSaveClick() {
@@ -18,16 +18,16 @@ function MoviesCard({ movie }) {
   }
   
   return (
-    <li className="movies-card">
+    <li className="movie-card">
       {
         pathname==="/movies" ?  
           <button className={saveButtonClassName} type="button" aria-label="Сохранить" onClick={handleSaveClick} /> :
-          <button className="movies-card__button movies-card__button_type_delete" type="button" aria-label="Удалить" />
+          <button className="movie-card__button movie-card__button_type_delete" type="button" aria-label="Удалить" />
       }
-      <img className="movies-card__image" src={`${MOVIES_BASE_URL}${movie.image.url}`} alt={movie.nameRU} />
-      <div className="movies-card__text-container">
-        <h2 className="movies-card__title">{movie.nameRU}</h2>
-        <p className="movies-card__duration">{formatTime(movie.duration)}</p>
+      <img className="movie-card__image" src={`${MOVIES_BASE_URL}${movie.image.url}`} alt={movie.nameRU} />
+      <div className="movie-card__text-container">
+        <h2 className="movie-card__title">{movie.nameRU}</h2>
+        <p className="movie-card__duration">{formatTime(movie.duration)}</p>
       </div>
     </li>
   );
